pyproject.toml
[project]
name = "gmail-bills-and-statements-watcher"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "firebase-admin>=6.8.0",
    "functions-framework>=3.8.3",
    "google-api-python-client>=2.170.0",
    "google-auth-httplib2>=0.2.0",
    "google-auth-oauthlib>=1.2.2",
    "google-cloud-firestore>=2.20.2",
    "google-cloud-secret-manager>=2.24.0",
    "loguru>=0.7.3",
    "mypy>=1.16.1",
    "pydantic-settings>=2.10.0",
    "pytest>=8.4.1",
    "python-dotenv>=1.1.0",
    "pyyaml>=6.0.2",
    "werkzeug>=3.1.3",
]
